#30: Small Cleanup (Sorting Donations & AdUnitId)

+++ android/DartsScorecard/domain/src/main/java/nl/entreco/domain/beta/donations/FetchDonationsUsecase.kt
@@ -13,7 +13,7 @@ class FetchDonationsUsecase @Inject constructor(private val billingRepository: B
 
     fun exec(done: (FetchDonationsResponse)->Unit, fail: (Throwable) -> Unit){
         onBackground({
-            val donations = billingRepository.fetchDonations().sortedBy { it.priceMicros }
+            val donations = billingRepository.fetchDonations().sortedBy { it.priceMicros.toLong() }
             val sorted = donations.sortedBy { it.price }
             onUi { done(FetchDonationsResponse(sorted)) }
         }, fail)
